Move EnvironFromName to cmd/juju/common.
