+++ cmd/juju/common/common.go
@@ -1,0 +1,10 @@
+// Copyright 2015 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package common
+
+import (
+	"github.com/juju/loggo"
+)
+
+var logger = loggo.GetLogger("juju.cmd.juju.common")
